simulationservice/requirements.txt
billiard==3.5.0.4
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
pytz==2018.5
vine==1.1.4
Werkzeug==0.14.1
Flask-Cors==3.0.6
decorator==4.3.0
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
idna==2.7
PyJWT==1.7.1
pycparser==2.19
gevent==1.3.7
StrawberryFields==0.23.0